requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
chromadb>=0.5.0
requests>=2.32.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
